Lists the downloads created by a user.

Lists the downloads created by a user.

occ_download_list(
  user = NULL,
  pwd = NULL,
  limit = 20,
  start = 0,
  curlopts = list()
)

Arguments

user
(character) User name within GBIF's website. Required. See Details.

pwd
(character) User password within GBIF's website. Required. See Details.

limit
(integer/numeric) Number of records to return. Default: 20, Max: 1000

start
(integer/numeric) Record number to start at. Default: 0

curlopts
list of named curl options passed on to HttpClient. see curl::curl_options for curl options

Value

a list with two slots:

meta: a single row data.frame with columns: offset, limit, endofrecords, count

results: a tibble with the nested data flattened, with many columns with the same request. prefix

Note

see downloads for an overview of GBIF downloads methods

See also

Examples

if (FALSE) {
occ_download_list(user="sckott")
occ_download_list(user="sckott", limit = 5)
occ_download_list(user="sckott", start = 21)
}