Defunct functions in rgbif

density_spplist(): service no longer provided

densitylist(): service no longer provided

gbifdata(): service no longer provided

gbifmap_dens(): service no longer provided

gbifmap_list(): service no longer provided

occurrencedensity(): service no longer provided

providers(): service no longer provided

resources(): service no longer provided

taxoncount(): service no longer provided

taxonget(): service no longer provided

taxonsearch(): service no longer provided

stylegeojson(): moving this functionality to spocc package, will be removed soon

togeojson(): moving this functionality to spocc package, will be removed soon

gist(): moving this functionality to spocc package, will be removed soon

occ_spellcheck(): GBIF has removed the spellCheck parameter from their API

Details

The above functions have been removed. See https://github.com/ropensci/rgbif and poke around the code if you want to find the old functions in previous versions of the package