Interface to the Global Biodiversity Information Facility API.

rgbif: A programmatic interface to the Web Service methods provided by the Global Biodiversity Information Facility.

Note

See many-values for discussion of how functions vary in how they accept values (single vs. many for the same HTTP request vs. many for different HTTP requests)

About

This package gives you access to data from GBIF https://www.gbif.org/ via their API.

A note about the old GBIF API

The old GBIF API is now defunct - that is, not available anymore. We used to have functions that worked with the old API, but those functions are now not available anymore because GBIF made the old API defunct.

Documentation for the GBIF API

summary https://www.gbif.org/developer/summary - Summary of the GBIF API

registry https://www.gbif.org/developer/registry - Metadata on datasets, and contributing organizations

species names https://www.gbif.org/developer/species - Species names and metadata

occurrences https://www.gbif.org/developer/occurrence - Occurrences

maps https://www.gbif.org/developer/maps - Maps - these APIs are not implemented in rgbif, and are meant more for intergration with web based maps.

Author

Scott Chamberlain

Karthik Ram

Dan Mcglinn

Vijay Barve