Lookup names in the GBIF backbone taxonomy.

Lookup names in the GBIF backbone taxonomy.

name_backbone(
  name,
  rank = NULL,
  kingdom = NULL,
  phylum = NULL,
  class = NULL,
  order = NULL,
  family = NULL,
  genus = NULL,
  strict = FALSE,
  verbose = FALSE,
  start = NULL,
  limit = 100,
  curlopts = list()
)

name_backbone_verbose(
  name,
  rank = NULL,
  kingdom = NULL,
  phylum = NULL,
  class = NULL,
  order = NULL,
  family = NULL,
  genus = NULL,
  strict = FALSE,
  start = NULL,
  limit = 100,
  curlopts = list()
)

Arguments

name
(character) Full scientific name potentially with authorship (required)

rank
(character) The rank given as our rank enum. (optional)

kingdom
(character) If provided default matching will also try to match against this if no direct match is found for the name alone. (optional)

phylum
(character) If provided default matching will also try to match against this if no direct match is found for the name alone. (optional)

class
(character) If provided default matching will also try to match against this if no direct match is found for the name alone. (optional)

order
(character) If provided default matching will also try to match against this if no direct match is found for the name alone. (optional)

family
(character) If provided default matching will also try to match against this if no direct match is found for the name alone. (optional)

genus
(character) If provided default matching will also try to match against this if no direct match is found for the name alone. (optional)

strict
(logical) If TRUE it (fuzzy) matches only the given name, but never a taxon in the upper classification (optional)

verbose
(logical) should the function give back more (less reliable) results. See function name_backbone_verbose()

start
Record number to start at. Default: 0. Use in combination with limit to page through results.

limit
Number of records to return. Default: 100. Maximum: 1000.

curlopts
list of named curl options passed on to HttpClient. see curl::curl_options for curl options

Value

For name_backbone, a data.frame for a single taxon with many columns. For name_backbone_verbose, a larger number of results in a data.frame the results of resulting from fuzzy matching. You will also get back your input name, rank, kingdom, phylum ect. as columns input_name, input_rank, input_kingdom ect. so you can check the results.

Details

If you don't get a match, GBIF gives back a data.frame with columns synonym, confidence, and matchType='NONE'.

References

https://www.gbif.org/developer/species#searching

Examples

if (FALSE) {
name_backbone(name='Helianthus annuus', kingdom='plants')
name_backbone(name='Helianthus', rank='genus', kingdom='plants')
name_backbone(name='Poa', rank='genus', family='Poaceae')

# Verbose - gives back alternatives
## Strictness
name_backbone_verbose(name='Poa', kingdom='plants',
  strict=FALSE)
name_backbone_verbose(name='Helianthus annuus', kingdom='plants',
  strict=TRUE)

# Non-existent name - returns list of lenght 3 stating no match
name_backbone(name='Aso')
name_backbone(name='Oenante')

# Pass on curl options
name_backbone(name='Oenante', curlopts = list(verbose=TRUE))
}