Retrieves the occurrence download metadata by its unique key.

Retrieves the occurrence download metadata by its unique key.

Usage

occ_download_meta(key, curlopts = list())

Arguments

key

A key generated from a request, like that from occ_download

curlopts

list of named curl options passed on to HttpClient. see curl::curl_options for curl options

Value

an object of class occ_download_meta, a list with slots for the download key, the DOI assigned to the download, license link, the request details you sent in the occ_download() request, and metadata about the size and date/time of the request

Note

see downloads for an overview of GBIF downloads methods

See also

Examples

if (FALSE) { # \dontrun{
occ_download_meta(key="0003983-140910143529206")
occ_download_meta("0000066-140928181241064")
} # }